// admin/backend/src/enums/data-routes.ts
export enum DataRoutes {
    artist = 'artist',
    genre = 'genre',
    genreTracks = 'genre-tracks',
    playlist = 'playlist',
    playlistTracks = 'playlist-tracks',
    track = 'track',
    user = 'user',
    userPlaylists = 'user-playlists',
    userTokens = 'user-tokens',
    userTracks = 'user-tracks',
}